requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0,<2.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
albumentations>=1.3.0
scikit-learn>=1.0.0
tensorboard>=2.8.0
wandb>=0.13.0
pycocotools>=2.0.6
onnx>=1.12.0
onnxruntime>=1.12.0
tensorrt>=8.4.0
polygraphy>=0.42.0
pyyaml>=6.0
hydra-core>=1.2.0
omegaconf>=2.2.0
timm>=0.6.0
transformers>=4.20.0
segmentation-models-pytorch>=0.3.0 